continuing work on v.10
	- Split SensorsCore.cpp into multiple files for easier maintenance and
		future growth for alternate sensors.
	- Renamed ESParaSite_ files to remove prefix.
	- Added routine to allow remotely setting of the Real Time Clock to UTC
		currently performed alongside GUI upload.
	- UPLOADER - added time setting function that sets the RTC to UTC
	- GUI - Removed code in home.js that converted time to UTC.
	- GUI - Renamed logo file to logo.png
