+++ VERSION
@@ -21,34 +21,41 @@
 
 /*
 v.10 - Kala-azar
-	- Moved GUI data feeds to be a single page with query strings
+	- Moved GUI data feeds to be a single URL with query strings
 	- Moved all HTTP_GET handlers that return JSON to their own cpp file
 	- Moved all HTTP Handlers for the web GUI to their own file.
 	- Moved all HTTP files under a unified header
 	- Moved all HTTP handlers for resetting counters to their own file.
 	- Added a GUI data feed for status page
-	- A kind gentleman contributed some of his time to rebuild the core 
+	- GUI - A kind gentleman contributed some of his time to rebuild the core 
 		of the GUI for us
-	- Added a Status Page.
 	- Moved all JSON handling functions to the DataToJson Namespaces
 		- Moved HistorytoJson to its own file for sake of code length.
 	- Moved all feeds to /guiFeed and renamed all to be consistent.
-	- Added Status Page to GUI showing the most critical information on one 
-		page.
-	- Added Diagnostics Page to the GUI which shows all information we know 
-		about the printer.
+	- GUI - Added Status Page to GUI showing the most critical information
+	 	on one page.
+	- GUI - Added Diagnostics Page to the GUI which shows all information 
+		we know about the printer.
 	- Changed Flash Layout on ESP12F to have a 3MB filesystem and 512K for 
 		OTA updates This update will format your flash and force you to 
 		reconfigure wifi.
 	- Removed ESP32-Targz in favor of a python GUI uploader.
-	- Created esps-uploader.py to load/update GUI onto ESParaSite
-	- GUI will now be uploaded/installed/updated via an external Python 
+	- UPLOADER -esps-uploader.py Created to load/update GUI onto ESParaSite
+	- GUI - will now be uploaded/installed/updated via an external Python 
 		Bscript.
 	- Renamed HTTPFile.cpp to HTTPHandleFile.cpp
-	- Added File listing and FS Info to Diagnostics page.
-	- Added debug options to GUI .js pages.  Pages can either pull 
+	- GUI - Added File listing and FS Info to Diagnostics page.
+	- GUI - Added debug options to GUI .js pages.  Pages can either pull 
 		from a http://esparasite.local (Needs CORS Plugin) or a local json/ 
 		subdirectory with debug=1 or debug=2 enabled.
+	- Split SensorsCore.cpp into multiple files for easier maintenance and 
+		future growth for alternate sensors.
+	- Renamed ESParaSite_ files to remove prefix.
+	- Added routine to allow remotely setting of the Real Time Clock to UTC
+		currently performed alongside GUI upload.
+	- UPLOADER - added time setting function that sets the RTC to UTC
+	- GUI - Removed code in home.js that converted time to UTC.
+	- GUI - Renamed logo file to logo.png
 
 v0.9 - Iodamoeba buetschlii - 2020-05-24
 	- Implemented locally served chart.js and jquery.js in order to start
