+++ web/esps-uploader.py
@@ -18,6 +18,9 @@ ESParaSite - GUI Upoloader v0.1
 
 import os
 import requests
+import math
+
+from datetime import datetime, timezone
 
 from os import listdir
 from os.path import isfile, join
@@ -29,8 +32,19 @@ filesDir = os.path.join(os.getcwd(), "gui")
 fileList = [f for f in listdir(filesDir) if isfile(join(filesDir, f))]
 
 uploadUrl = "http://" + espsMdns + "/upload"
-
+"""
 for file in fileList:
     filePath = os.path.join(filesDir, file)
     with open(filePath, 'rb') as f:
         r = requests.post(uploadUrl, files={filePath: f})
+
+"""
+# seconds from epoch:
+now_utc = datetime.now(timezone.utc).timestamp()
+now_utc_sec = math.floor(now_utc)
+print(now_utc_sec)
+
+uploadUrl = "http://" + espsMdns + "/setRtc"
+payload = {'TimeStamp': now_utc_sec}
+
+r = requests.post(uploadUrl, params=payload)
